Fix MediaDevices-produceCropId WPT test
This patch fixes the currently broken produceCropId test. Due to an
issue with the test harness parsing an iframe, this test set is not
picked up and ran. The produceCropId API has also been finalized and
removed from the Element API, so some additional changes to fix the
actual tests as well.

Bug: 1314679
Change-Id: Ic967c6b2c2651ac9391078897219ed337a033513