Adjust when simplified layout can be performed
We could end up in a state where we retrieved a cached *measure*
result, and a cache status of kNeedsSimplifiedLayout. However, if the
box had previously fragmented, it wouldn't have a cached layout result
(and its children might not either). We should avoid simplified
layout in this case since simplified layout relies on children having
a cached layout result.

Bug: 1321393
Change-Id: I9748c75cebbdefc7946e27c40bc8f6024cf305af